+++ docs/sidebar.md
@@ -3,9 +3,10 @@
 [git repo](https://github.com/wavelab/libwave)
 
 ## API
-- [kinematics](#api/kinematics/kinematics)
-    - [rotation](#api/kinematics/rotation)
 - [estimation](#api/estimation/estimation)
+- [geometry](#api/geometry/geometry)
+    - [rotation](#api/geometry/rotation)
+- [matching](#api/matching/matching)
 - [optimization](#api/optimization/optimization)
 - [utils](#api/utils/utils)
