+++ wave_geometry/include/wave/geometry/rotation.hpp
