+++ wave_geometry/src/rotation.cpp
@@ -1,4 +1,4 @@
-#include "wave/kinematics/rotation.hpp"
+#include "wave/geometry/rotation.hpp"
 #include "wave/utils/utils.hpp"
 
 namespace wave {
